fix: address review feedback (attempt 2) — Add Tavily as configurable agent search tool
