+++ docs/agent/configuration.md
@@ -93,8 +93,8 @@ The following shortcut strings load tools directly. Passing a Tool instance is a
 | http.*      | HTTP Path to a [Model Context Protocol (MCP)](https://modelcontextprotocol.io/docs/getting-started/intro) server |
 | python      | Runs a Python action                                      |
 | read        | Reads file or url content, supports text extraction       |
-| todowrite   | Generates a task list to organize complex tasks           |
 | tavily      | Runs a web search using the [Tavily API](https://tavily.com). Requires `TAVILY_API_KEY` env var and `tavily-python` package (install via `pip install txtai[agent-tavily]`) |
+| todowrite   | Generates a task list to organize complex tasks           |
 | websearch   | Runs a websearch using the built-in websearch tool        |
 | webview     | Extracts content from a web page. Alias for `read` tool   |
 | write       | Writes content to file                                    |
